change: update examples
