+++ examples/uitest/demo_android_feed_swipe.json
